backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djoser==2.2.3
Pillow==11.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
gunicorn==23.0.0
django-filter==24.3
